handle get and create cart api

+++ lib/models/cart_product_model.dart
@@ -2,8 +2,9 @@ import 'package:pharma_app/models/product_model.dart';
 
 class CartProduct {
   final Product product;
-  final double quantity;
+  final int quantity;
 
   CartProduct(this.product, this.quantity);
-  factory CartProduct.fromJson(Map<String, dynamic> json) => CartProduct.fromJson(json);
+  factory CartProduct.fromJson(Map<String, dynamic> json) =>
+      CartProduct.fromJson(json);
 }
